+++ packages/quantic/cypress/e2e/default-1/tab-bar/tab-bar.cypress.ts
@@ -10,6 +10,11 @@ import {
 } from '../../../page-objects/use-case';
 import {performSearch} from '../../../page-objects/actions/action-perform-search';
 
+interface TabBarOptions {
+  lightTheme: boolean;
+  useCase: string;
+}
+
 describe('quantic-tab-bar', () => {
   const pageUrl = 's/quantic-tab-bar';
   const extraSmallViewportWidth = 200;
@@ -21,16 +26,19 @@ describe('quantic-tab-bar', () => {
   const TAB_3 = 'Tab 3';
   const TAB_4 = 'Tab 4';
 
-  function visitPage(useCase: string, waitForSearch = true) {
+  function visitPage(
+    options: Partial<TabBarOptions> = {},
+    waitForSearch = true
+  ) {
     interceptSearch();
     cy.visit(pageUrl);
-    configure({useCase: useCase});
-    if (useCase === useCaseEnum.insight) {
+    configure(options);
+    if (options.useCase === useCaseEnum.insight) {
       InsightInterfaceExpect.isInitialized();
       performSearch();
     }
     if (waitForSearch) {
-      cy.wait(getAlias(useCase));
+      cy.wait(getAlias(options.useCase));
     }
   }
 
@@ -38,10 +46,11 @@ describe('quantic-tab-bar', () => {
     describe(param.label, () => {
       describe("when the container's width can fit all the tabs", () => {
         it('should display all the tabs without displaying the dropdown list', () => {
-          visitPage(param.useCase);
+          visitPage({useCase: param.useCase});
 
           scope('when loading the page', () => {
             Expect.displayTabs(true);
+            Expect.displayWithLightTheme(false);
             Expect.displayedTabsEqual([TAB_1, TAB_2, TAB_3, TAB_4]);
             Expect.activeTabContains(TAB_1);
             Expect.displayMoreButton(false);
@@ -58,10 +67,11 @@ describe('quantic-tab-bar', () => {
       describe('when the viewport is resized to a medium width', () => {
         it('should display only the first two tabs, the other tabs should be displayed in the dropdown list', () => {
           cy.viewport(mediumViewportWidth, 900);
-          visitPage(param.useCase);
+          visitPage({useCase: param.useCase});
 
           scope('when loading the page', () => {
             Expect.displayTabs(true);
+            Expect.displayWithLightTheme(false);
             Expect.displayedTabsEqual([TAB_1, TAB_2]);
             Expect.tabsInDropdownEqual([TAB_3, TAB_4]);
             Expect.activeTabContains(TAB_1);
@@ -91,10 +101,11 @@ describe('quantic-tab-bar', () => {
       describe('when the viewport is resized to a small width', () => {
         it('should display only the selected tab, the other tabs should be displayed in the dropdown list', () => {
           cy.viewport(smallViewportWidth, 900);
-          visitPage(param.useCase);
+          visitPage({useCase: param.useCase});
 
           scope('when loading the page', () => {
             Expect.displayTabs(true);
+            Expect.displayWithLightTheme(false);
             Expect.displayedTabsEqual([TAB_1]);
             Expect.tabsInDropdownEqual([TAB_2, TAB_3, TAB_4]);
             Expect.activeTabContains(TAB_1);
@@ -119,10 +130,11 @@ describe('quantic-tab-bar', () => {
       describe('when the viewport is resized to an extra small width', () => {
         it('should display only the selected tab, the other tabs should be displayed in the dropdown list and the more button should be shrunk', () => {
           cy.viewport(extraSmallViewportWidth, 900);
-          visitPage(param.useCase);
+          visitPage({useCase: param.useCase});
 
           scope('when loading the page', () => {
             Expect.displayTabs(true);
+            Expect.displayWithLightTheme(false);
             Expect.displayedTabsEqual([TAB_1]);
             Expect.tabsInDropdownEqual([TAB_2, TAB_3, TAB_4]);
             Expect.activeTabContains(TAB_1);
@@ -147,7 +159,7 @@ describe('quantic-tab-bar', () => {
       describe('when the dropdown loses focus', () => {
         it('should automatically close the dropdown list', () => {
           cy.viewport(mediumViewportWidth, 900);
-          visitPage(param.useCase);
+          visitPage({useCase: param.useCase});
 
           scope('when opening the dropdown list', () => {
             Actions.openDropdown();
@@ -164,7 +176,7 @@ describe('quantic-tab-bar', () => {
       describe('when a tab containing the same expression as another tab is selected from the dropdown list', () => {
         it('should correctly select the tab', () => {
           cy.viewport(extraSmallViewportWidth, 900);
-          visitPage(param.useCase);
+          visitPage({useCase: param.useCase});
 
           scope('when loading the page', () => {
             Expect.displayedTabsEqual([TAB_1]);
@@ -178,6 +190,17 @@ describe('quantic-tab-bar', () => {
             Actions.selectTabFromDropdown(TAB_4);
             Expect.displayedTabsEqual([TAB_4]);
             Expect.tabsInDropdownEqual([TAB_1, TAB_2, TAB_3]);
+          });
+        });
+      });
+
+      describe('when the light theme property is set to true', () => {
+        it('should display the component with the light theme styles', () => {
+          visitPage({useCase: param.useCase, lightTheme: true});
+
+          scope('when loading the page', () => {
+            Expect.displayTabs(true);
+            Expect.displayWithLightTheme(true);
           });
         });
       });