+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.111.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.111.0...@coveo/headless@1.111.1) (2022-11-10)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
+# [1.111.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.110.4...@coveo/headless@1.111.0) (2022-11-10)
+
+
+### Features
+
+* **headless:** added insight query suggest support ([#2536](https://github.com/coveo/ui-kit/issues/2536)) ([01f936b](https://github.com/coveo/ui-kit/commit/01f936bc214edf7bd724f241b18ecfb3148b8108))
+
+
+
+
+
+## [1.110.4](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.110.3...@coveo/headless@1.110.4) (2022-11-08)
+
+
+### Bug Fixes
+
+* **headless:** folding cq should support spaces ([#2527](https://github.com/coveo/ui-kit/issues/2527)) ([28d0392](https://github.com/coveo/ui-kit/commit/28d0392a8fae5a97dd53aa70215e061605f61dfc))
+
+
+
+
+
+## [1.110.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.110.2...@coveo/headless@1.110.3) (2022-11-04)
+
+**Note:** Version bump only for package @coveo/headless
+
+
+
+
+
 ## [1.110.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.110.1...@coveo/headless@1.110.2) (2022-11-02)